refactor(infra): isolate approval type contracts
