+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -70,7 +70,7 @@ ceacad83db01c66e7be6aa21a291597020f13f737b697690eae7d47098e6499a  module/gateway
 182dc685f2103ff66c1a4839a48f4f40d2eeb0070cd74e449b47aacc4e6f1c22  module/hook-runtime
 f6e3c44e7d1090a97aca554a3c247219b8de78b3cb4399cac5efde8a0a6c1156  module/inbound-envelope
 f1fa557a1bf0d50c8943433348404bfa5a2e8792199164507c0904f5d288a48d  module/inbound-reply-dispatch
-3e4d83cd50db349362fe7aa3ec9dfce779243c2fecff243c9fb027061856c8f6  module/infra-runtime
+479ff50fc9f778d0062c3742ee2070f3c2722d8fedc45a3ea25a97da5676a14c  module/infra-runtime
 2e717cccb3db127aed0287d4ea14c41a8e64e46d60c728638153e31e2fb0d296  module/ingress-effect-once
 7c0bc9455678102a68cc80796ca04ba16f1fe9980335de9d58f82b830aef6876  module/interactive-runtime
 9dd66baf2def46386ad4706380e57f9068fa9bb3878be2d9ba961ec2f46d3d87  module/json-store
